Rename test files and update import paths for consistency
