+++ src/components/__tests__/Loader.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import { describe, it, expect } from 'vitest'
-import Loader from './Loader'
+import Loader from '../Loader'
 
 describe('Loader Component', () => {
   it('renders a LinearProgress bar', () => {
