+++ src/components/__tests__/Missing.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import { describe, it, expect } from 'vitest'
-import Missing from './Missing'
+import Missing from '../Missing'
 
 describe('Missing Component', () => {
   it('renders the 404 image with correct alt text', () => {
